+++ app/src/main/java/com/benhagy/gamefinder3/domain/repository/GamefinderRepository.kt
@@ -36,7 +36,7 @@ interface GamefinderRepository {
 
     suspend fun upsertUserNote(note: String, id: Int)
 
-    suspend fun upsertUserRating(rating: Double, id: Int)
+    suspend fun upsertUserRating(rating: Float, id: Int)
 
 
 }
